+++ shiki_organizer/cli/repository.py
@@ -100,14 +100,20 @@ def delete(id):
 def pull(github_token):
     def process_issue(issues, repository, is_closed=False):
         for i in issues:
+            created = False
             issue, _ = Issue.get_or_create(id=i.number, repository=repository)
             if not issue.task:
                 task = Task.create(
                     description=i.title,
                 )
                 issue.task = task
+                print(f"Created task '{task.description}'.")
+                created = True
             task = issue.task
+            old_description = task.description if task.description else None
             task.description = f"{i.title} #{issue.id}"
+            if not created and old_description and task.description != old_description:
+                print(f"Rename task '{old_description}' to '{task.description}'.")
             task.parent = repository.parent
             task.archived = is_closed
             task.save()